PY3 migration blink-web-tests: Fix race/deadlock.
There is a race/deadlock between worker exits and message queue. This is mostly seen during the retries on linux. Some workers may finish execution before the manager processed the message queue.
This ends up in a deadlock with parent getting stuck in the queue.

In Linux by default processes are started using fork and mac and windows have spawn by default.
while fork is faster to start, spawn is safer. This race was seen only on linux not on mac.

Changing the process start method to 'spawn' fixes the problem in linux as well.

Bug: 1197679

Change-Id: Ifebeb1194e272aa60cf0a59ab3ef7d5ec0e0a772
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3015250
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Preethi Mohan <preethim@google.com>
Cr-Commit-Position: refs/heads/master@{#900235}
NOKEYCHECK=True
GitOrigin-RevId: 560be8eeb0e698c25b991eb67ac626bd2f189f58