[mathml] Invalidation after dynamic change of mfrac@linethickness
The <mfrac>'s linethickness attribute [1] is mapped to an internal style
property. It affects the block size of the painted fraction bar as well
as the block offsets of the numerator/denominator. This CL adds some
logic to MathMLFractionElement in order to perform the required
invalidations.

[1] https://w3c.github.io/mathml-core/#dfn-linethickness

Bug: 6606
Change-Id: Ie3ad76a9d36be82e120819284ee395da183af3bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3016649
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#900465}
NOKEYCHECK=True
GitOrigin-RevId: 76230ef27d59920752b4a192b486aebd96b97d42