usb: Use ReadOnlyBuffer in the rest of the Mojo interface
This change replaces the rest of the instances of array<uint8> in Mojo
method arguments and replies with the typemapped ReadOnlyBuffer struct.
This avoids an unnecessary copy of the transferred data on both the
sending and receiving ends of the Mojo pipe.

Bug: 743121
Change-Id: I1ac823790341eefc60ff3c1ae0e2a82587e4c51c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2919266
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Chris Mumford <cmumford@google.com>
Cr-Commit-Position: refs/heads/master@{#900787}
NOKEYCHECK=True
GitOrigin-RevId: e45c567c11d4be12f8557f7314ce3149491c55d3