[TableNG] Fix percentage relative-positioning.
Previously we'd set a fixed block-size on various table parts. The way
that %-based relative positioning would work is to use this fixed
block-size to resolve %-ages against.

We weren't really using this fixed block-size however - as we'd just set
the final block-size as the sum of our children.

Instead of setting the fixed block-size, just use the intrinsic
block-size. This requires a change to the simplified layout algorithm
to "ignore" one of the DCHECKs we perform.

This behaviour now resolves %-relpos against any *specified* size on
the parent. This matches Firefox with the exception of test
position-relative-013.html .

This behaviour seems reasonable - however if we want to restrict this
even more, we can set the IsInitialBlockSizeIndefinite to true (then
all %-respos won't resolve).

Bug: 1227884
Change-Id: Id036a52b954604f1d26528a557053e46ea5aa0be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3021544
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Aleks Totic <atotic@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#900823}
NOKEYCHECK=True
GitOrigin-RevId: 3572711a2c6af39177d5a6f2da8090960b402a85