Propagate anonymous attribute to the browser
This CL adds logic for propagating the (experimental) iframe
'anonymous' attribute to the browser process. In particular, this CL
adds a boolean `anonymous_` parameter to the FrameTreeNode, the
NavigationRequest and the RenderFrameHost, with the following
meanings:
- A FrameTreeNode is anonymous if either the corresponding iframe has
  the 'anonymous' attribute, or if its parent is an anonymous
  RenderFrameHost.
- A NavigationRequest is anonymous if it will commit into an anonymous
  FrameTreeNode.
- A RenderFrameHost is anonymous if it was committed by an anonymous
  NavigationRequest.

See the design document for more details:
https://docs.google.com/document/d/1f8lDmVCPc4CA8zCIlVtVcYEHStFJOEimqltB1z6_W1U/edit

Note: the 'anonymous' attribute is currently parsed only when the
blink runtime feature "AnonymousIframe" is enabled. Hence this whole
code is a no-op unless that feature flag is set.

Bug: 1226469
Change-Id: Ifbb097dfe21afde0ce4f8f5779f3660d8df621e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3002073
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/master@{#900899}
NOKEYCHECK=True
GitOrigin-RevId: 5abc8de2532dc64c13419a8948acc401024b58a3