prerender: Pass is_prerendering to RenderView instead of nav commit.
Before this CL, the browser only tells the renderer that a document
is being prerendered as part of the CommitNavigation IPC. But some
documents in the renderer are created before CommitNavigation or do
not go through it. Therefore these documents do not know that they
are being prerendered.

This CL fixes the issue by moving the is_prerendering bit from
CommitNavigation to CreateRenderView. This ensures that RenderViews
understand whether they are being prerendered, and any document that
is created in that RenderView/Page starts out with the desired state.

We also add an Activate IPC to RenderView, which means that documents
created after that IPC arrives start out with the desired state.
This is redundant with the existing per-frame IPC, which a subsequent
CL will remove. The per-frame IPC is kept only as an intermediate
implementation step.

This is the result of a discussion on navigation-dev[1]. A summary
is of the design is available at [2].

[1] https://groups.google.com/a/chromium.org/g/navigation-dev/c/v3T8NfWPTg8/m/brmF98nKAAAJ
[2] https://docs.google.com/document/d/1S8_ilRC_TkDa4Wz3PvBPYhhqIuGNHDX0DKsvjMFotKM/edit?usp=sharing

Bug: 1185965
Change-Id: I4876655615782de0d6645853f12a598637675d52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3017926
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Owners-Override: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901189}
NOKEYCHECK=True
GitOrigin-RevId: 84e9a9b442fb85b9df2bc45aca3d47913285d4a5