Add regression tests for origin trials enabled after window.open
There were no tests to ensure that trials are enabled in a document that
was navigated via window.open.

This CL adds tests for that scenario, using a new test-only attribute
added to the Window interface. This can be extended in the future with
a test-only attribute on other globals (i.e. ServiceWorker).

Also adds tests specifically for storageFoundation, which is the trial
where the problem was discovered.

Bug: 1206219
Change-Id: If533a259f62f73a5d89802ca7c542b2368cc234e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2910277
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Rodney Ding <rodneyding@google.com>
Commit-Queue: Jason Chase <chasej@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901257}
NOKEYCHECK=True
GitOrigin-RevId: f64cc6d06043378fecdd1372d153da2a04c91ff5