Fix LayoutText::PlainText() to handle floating ::first-letter
LayoutText::PlainText() is supposed to return the full text of the
text node, even if it's called on the remaining text part of a
::first-letter.

crrev.com/c/2654312 switched the function to use NGOffsetMapping, and
introduced a regression that it fails to collect floating first letter.
This is because floating first letter and its remaining text are in
different inline formatting contexts and use different NGOffsetMapping,
and that patch collects text only from the remaining text's mapping.

This patch adds special handing to this case so that floating first
letter text is also collected.

Bug: 1227955
Change-Id: Ica1bc36f8ffad9d56ffb45c10f1e0314ae505dff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3022570
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901270}
NOKEYCHECK=True
GitOrigin-RevId: c4a02aba63c97f06a0956ebae4eee3a37c3130e9