[storage] Add nonce to StorageKey
This CL adds an optional nonce member to blink::StorageKey and
blink::BlinkStorageKey, and updates the mojo traits and the unittests.

The nonce's purpose is to further partition two contexts which have
the same origin. In particular, it will be used for implementing
anonymous iframes.

No code sets the nonce at the moment, and the StorageKey without nonce
behaves exactly as before, so this CL does not change current
behaviour.

Bug: 1199077,1226469
Change-Id: I7c3a1989d827adbdc2f61f40c9c00d380b288ad2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3015202
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901454}
NOKEYCHECK=True
GitOrigin-RevId: 2cfef283d649c8c7925345542831ccf86ca9919f