EventTargetData: Avoid inlining a GarbageCollected object
Use a combination of a mixin, private inheritance, and a separate
object to allocate to avoid the GC_PLUGIN_IGNORE.

In addition, having a proper mixin interface actually allows for the
EventTargetData reference to actually escape onto the stack or heap,
which would otherwise be hard to check and enforce.

The only trade off is that this approach minmally regressed tracing
performance of EventTargetDataMap because the GC has to lookup the
header of a mixin object.

Bug: 513199, 1229149
Change-Id: I304655f4805810c38304b49526fb01b6d2eb87d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3026825
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901470}
NOKEYCHECK=True
GitOrigin-RevId: 0a778ea3c2719dc76b3331db644abb66210a785e