[LayoutNG] OOF static position with conflicting direction
We were always using the containing block's writing mode when
converting the static position because we want the offset to be
relative to the CB. However, this resulted in an incorrect offset
when the multicol had a different direction. Instead of always
using the CB's writing mode when converting the static position, use
the current fragment's writing mode.

The added tests are based on their non-multicol counterparts. Some
of the tests passed previously if there was no writing mode mismatch.

Bug: 1228746
Change-Id: I2027ccd7c384a75f54d8dc5833de775f3ab7e58c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3025848
Commit-Queue: Alison Maher <almaher@microsoft.com>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901570}
NOKEYCHECK=True
GitOrigin-RevId: 03a820b4eb5ef046113b88e339974dbb3e623dc2