Canvas context type 'gpupresent' -> 'webgpu'
Adds support for the 'webgpu' canvas context type and deprecates
'gpupresent' to match the latest spec changes.

Also renames GPUSwapChainDescriptor to GPUCanvasConfiguration, since
that was also changed in the spec for consistency. No deprecation period
is needed for that change since dictionary names are not visible to
developers.

Bug: 1229274
Change-Id: Iec1ce15a4eb16c186c19b9291f150611fd6220bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3027671
Reviewed-by: Austin Eng <enga@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901690}
NOKEYCHECK=True
GitOrigin-RevId: 49e0293fd261fde760170473ec0158aed6b2dcaf