[LayoutNG] OOF inline containing block writing-mode conversions
When we create the containing block rect based on the geometry
calculated in InlineContainingBlockUtils, we convert everything
against the entire containing block, including the block size from
all fragments. However, the start/end rects that are calculated
in InlineContainingBlockUtils are relative to a single containing
block fragment.

As a result, we can end up with incorrect offsets when performing
writing mode conversions on the containing block rect. To fix this,
convert the item offsets in InlineContainingBlockUtils such that they
are relative to the full containing block instead.

In addition to the writing-mode conversion improvements, it was
also observed that we weren't retrieving the previous containing
block break token in InlineContainingBlockUtils::
ComputeInlineContainerGeometryForFragmentainer() when there were
no split inlines. The reason being that we would continue to the
next CB fragment before updating the previous break token. This
logic is fixed by this CL, as well.

Bug: 1227115
Change-Id: I1039ef0c93ef928891c3b581399b3d3f1427d6b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3017280
Commit-Queue: Alison Maher <almaher@microsoft.com>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901795}
NOKEYCHECK=True
GitOrigin-RevId: 3f2f450312882e73504f82df00b889384ffe8c97