Use the correct origin for determining COOP.
When checking COOP, use the final origin of the document instead
of the one derived from its URL. It means sandbox_flags and
opaque origin are taken into account.

Problem:, there was the following dependencies:
- origin depends on sandbox
- sandbox depends on CSP
- CSP depends on the final PolicyContainer
- The final PolicyContainer depends on the retrieved COOP policy.

So, introducing:
- COOP depends on origin.
would form a cycle.

It turns out determining the COOP policy from the headers doesn't
require to know the final origin of the document. Determining if a
context is secure from the specification is only about checking the
document's URL instead of its origin.

This lead to COOP logic being split in half:
1. Retrieve the COOP policy from the headers.
2. Compute the final sandbox flags, origin, PolicyContainer, etc...
3. Enforce COOP with the final origin.

Bug: 1063518
Change-Id: Iaa53929bca60762364c8cc4f08db55039c6e74d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3021122
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901994}
NOKEYCHECK=True
GitOrigin-RevId: 64e31df42d36e431df00423d88aef876dbc905ff