Add partition_key to CanonicalCookie.
This CL adds an absl::optional<net::SchemefulSite> field to
CanonicalCookie which stores the cookie's partition key if the cookie
was set with the Partitioned attribute.

Any cookie that was set without the Partitioned attribute will have
absl::nullopt as a partition key.

This CL also adds methods to CanonicalCookie to return:
- IsPartitioned: if the cookie was set with the Partitioned attribute.
- PartitionKey: the cookie's partition key.

Bug: 1225444
Change-Id: I9c0fa52a28aeedaa631c438e2d1eb1ae7383bd57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3012034
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Commit-Queue: Dylan Cutler <dylancutler@google.com>
Cr-Commit-Position: refs/heads/master@{#902527}
NOKEYCHECK=True
GitOrigin-RevId: 01ca8e3a0bdfbf3e51dbe79af95725df64489553