Content with animating opacity should not be considered invisible
If an opacity animation is running on the compositor, the main thread
will not get a chance to update it if the opacity changes from zero to
non-zero. Therefore, we cannot skip the painting of content with an
active opacity animation. This regressed with https://crrev.com/897409.

This approach will also improve the microsoft_performance perf test
(live version at [1]) which has content with animating opacity.
CompositeAfterPaint regressed this benchmark because it spreads out the
raster of animating layers throughout the benchmark rather than doing
them all up-front.

[1] https://testdrive-archive.azurewebsites.net/Performance/RoboHornetPro/Default.html

Bug: 1226180, 1230093
Change-Id: I197a0592cdc776d9810d83350f8052c73b9e4de6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3034488
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#902611}
NOKEYCHECK=True
GitOrigin-RevId: fb417b078590f38e8cb33009fc6c9415e8ed5fb4