Fix flaky test layout-instability/input-timestamp.html
layout-instability/input-timestamp.html test is flaky because we are
waiting for 500ms to change the layout after a click event. The click
event should report a layout shift entry to PerformanceObserver with
hadRecentInput to be true, the layout shift of the div will report
another layout shift entry to PerformanceObserver with hadRecentInput to
be false. 500ms is the interval of the excluding input window, so if we
just set the timeout to be 500ms, the test is flaky, because the click
event may not executed right away. I change the time delay to be 600ms
to reduce the flakiness.

Bug: 1221422
Change-Id: Ib960d98fe7ac7d0806a7152052465e9188afb2a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3032268
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#902699}
NOKEYCHECK=True
GitOrigin-RevId: aa3b63cc92dfd6c12ed05342b630be5efcf3bab0