[BreakoutBox] Add Frame Limiter
This CL adds a subsystem to limit frames open by
MediaStreamTrackProcessors for capturers that may be subject to a limit
in the number of in-flight open frames.

At this time, this applies to camera, screen and window video capturers.
The limit is enforced per renderer process and is such that a single
renderer process cannot inferfere with another process also opening
frames. The actual limits vary depending on platform and capturer type.

As part of this work, explicit client locking and locked operations have
been added to FrameQueue, so that a combination of Push/Pop/Peek
operations can be performed while the queue is locked when deciding what
to do when a new frame arrives to a FrameQueueUnderlyingSource.

A base::Feature is added in order to make it easy to disable the limiter
codepath if necessary.

Bug: 1227946
Change-Id: Ie44b84c96bea0d7a5a20f7494aeecd15157a044c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3033082
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#902716}
NOKEYCHECK=True
GitOrigin-RevId: 503880f432726f96cf72759aae2c7d0fab1b8edc