[cleanup] Use IsInitialBlockSizeIsIndefinite for table-cells.
Previously we had a lot of manual logic to set the %-resolution-size
to indefinite for table-cells in the "measure" phase.

We how have IsInitialBlockSizeIsIndefinite! This directly maps onto what
we were doing previously.

This patch sets IsInitialBlockSizeIsIndefinite to true or table-cells
in the measure phase, and removes all the special logic.

We can also now remove the special caching logic for table cells as
we can now rely on the standard indefinite detection.

Change-Id: I8217b121da17362c6932226dae0461b25a9ec5ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3025075
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: David Grogan <dgrogan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#902791}
NOKEYCHECK=True
GitOrigin-RevId: 3cfaa26f262bcfd7ac0182f1da165ec42e2fe13f