[ua-ch] Remove Enterprise Policy for User Agent Client Hints
As noted in components/policy/resources/policy_templates.json, this
enterprise policy was intended for a short-term window to allow
enterprises to adapt to the new User Agent Client Hints, and was slated
for removal in Chrome 94.

The policy was removed following these instructions:
https://chromium.googlesource.com/chromium/src/+/main/docs/enterprise/add_new_policy.md#removing-support-for-a-policy.

The prefs were removed following these instructions:
https://chromium.googlesource.com/chromium/src/+/main/chrome/browser/prefs/README.md#deleting-an-old-pref.

The CL that added this enterprise policy was crrev.com/c/2248426.

Bug: 1217444
Change-Id: I8214cf509368ecfd8a9fb23d0c391b1710d18a5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3035821
Commit-Queue: Ali Beyad <abeyad@chromium.org>
Reviewed-by: Anqing Zhao <anqing@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Aaron Tagliaboschi <aarontag@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#903051}
NOKEYCHECK=True
GitOrigin-RevId: 0a4a3d71c111600c538d59528a0a21e70d966e1b