Modify parser yielding heuristic
This patch modifies the heuristic used to set the budget for the
synchronous budgeted HTML parser. It was previously a fixed (but Finch-
configurable) value of 250 tokens. It is now non-Finch-configurable,
and starts at 250 tokens. But after a total of three yields, the budget
is expanded to a very large number.

Bug: 1041006
Change-Id: Ibe670bcf5a66eab2897ce452ca220b5e10d488fa
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2983494
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/master@{#903116}
NOKEYCHECK=True
GitOrigin-RevId: 6662fe0e891cc25fbc8325b2ed7b10d0e9b43625