Roll the Android support library to a version with sources
I honestly have no idea if this is rolling to a newer or older
version, but the jars we pulled from maven had no sources or git hashes
associated with them so I can't compare the versions.

This version is pulled from
https://android.googlesource.com/platform/frameworks/testing/+/refs/heads/android-support-test
which will at least allow us to understand what the library is doing
since it's totally undocumented these days. It'll also allow us to
modify this library and fix bugs in it at least until we can migrate
to androidx (this library was deprecated in 2018(?)) as it is no longer
under development or supported.

Bug: 1230157
Change-Id: I76e55a696c363398eda04be09a41ae877c2596e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3035542
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Haiyang Pan <hypan@google.com>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#903590}
NOKEYCHECK=True
GitOrigin-RevId: e1c26ab5af72cfe8c6ebec4891fce736fc9589ef