[counter-style] Reorganize the legacy Armenian algorithm with a small fix
This CL reorganizes the implementation of the 'lower-armenian' and
'upper-armenian' predefined counter styles, by making it more compact
with the callers and using StringBuilder instead of character array.

It also fixes a small bug that counter value 0 may be incorrectly passed
to the main algorithm if we extend the counter style and expand its
range.

Bug: 1228531
Change-Id: I6512c3330485123c0eb5ef82fc475df221f8be48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3035797
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#903623}
NOKEYCHECK=True
GitOrigin-RevId: f48ea952945b41e459d92eff67905780666a8c78