[GridNG] Re-compute grid with a forced initial block-size indefinite.
Flexbox is able to force an item (the grid in this case) to have an
indefinite initial block-size. (See testcase for exact scenario).

Grid will typically re-compute the grid with the final used block-size.
The current behaviour might have a grid sized at 100px, but a single
auto track sized at 0px.

The complexity here is that we can't ignore indefinite initial
block-size as we need to compute the IntrinsicBlockSize correctly, as
flexbox will use this value in its algorithm.

The fix is to check when we have this flag set, and if this block-size
*is* definite (ignoring this flag) re-compute the grid in the outer
2nd-pass.

Bug: 1055258
Change-Id: I469ecc26c32ddc13a7285b8cb749eef17a317ec9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2994864
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Reviewed-by: David Grogan <dgrogan@chromium.org>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#904057}
NOKEYCHECK=True
GitOrigin-RevId: f1a24875232bd70605260eea7d13a7ea263ee579