binding: Improve V8 value conversions at js_event_handler.cc
This patch replaces the usage of old API (ScriptValue::From)
with the new API (ToV8Traits<T>::ToV8).

The root cause of crbug.com/1231434 is yet unknown though,
this patch could fix the issue (or help the investigation)
because this patch optimizes the conversion that crashes
and makes it almost no-op.

Even if this patch does nothing for the bug, this is a minor
improvement of the code base.

Bug: 1231434
Change-Id: I704a53bb1570f987681060c9483abff9e3fcae95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3043802
Auto-Submit: Yuki Shiino <yukishiino@chromium.org>
Commit-Queue: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#904127}
NOKEYCHECK=True
GitOrigin-RevId: cd3e88cc0f6256bb5ef7dc967e8c6ce537971175