Remove line-height related clamping
Currently line-height computed values that are fed into layout are
clamped, which prevents sub-pixel lines from being laid out. We've seen
two main issues reported with this - scripts that perform their own
calculations with the specified line-height number accumulate error, and
inconsistencies where device-scale-factor is not a whole number. In the
latter, use-zoom-for-dsf applies the scale factor on the computed
value, which is then floored, which leads to different layouts (from
script perspective) when moving between monitors with disparate DSF.

Removing this clamping results in behavior that we see in Gecko; landing
this shortly after the branch point to give us as much time as possible
to see whether this change is web-compatible.

Bug: 1023654

Change-Id: Ifc18457699f25c83037e1047acb3a47fa6a87836
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2824938
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#904217}
NOKEYCHECK=True
GitOrigin-RevId: 4f8ac4264908b7717de500800688c13c028c0831