Tint accent-color for hover and press
When hovering and pressing clickable form controls, different shades of
blue are used to render the control to show feedback.
This patch replicates this behavior when accent-color is used by
applying the same difference in luminance to the accent-color when
controls are hovered and pressed.

Fixed: 1228930
Change-Id: I48891273004972a6eed89aac645518e326c28dc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3043360
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/master@{#904537}
NOKEYCHECK=True
GitOrigin-RevId: 0de4a2b607fc234a85ee8d88c839dcb9173f480b