[SPC] Check UVPAA before starting SPC enrollment
Before this patch, the isUVPAA() check of SPC's makeCredential()
happened after we showed a SPC enrollment UI. This was incorrect.

After this patch, the isUVPAA() check of SPC's makeCredential()
will happen before we show a SPC enrollment UI.

Note that this CL also non-trivially changes the setup of the
spc_browsertest. The introduction of isUVPAA() in makeCredential()
failed all the tests because isUVPAA() returned false when not
manipulated and led to the "no-UVPAA" error. Since it's hard to mock
isUVPAA, this CL uses a SPCDebug flag to bypass the error.

Bug: 1230874
Change-Id: I668135bdb44cc6952a447f4abd5e40ccbc0f5475
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3042202
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#904544}
NOKEYCHECK=True
GitOrigin-RevId: 10a1360bfbed628c33dc7e002b43bc2c3e38828f