Return unsafe_agent_cluster_id from BlobURLStore.Resolve*
In order to track potential breakage if/when the blob store is
partitioned we need to track the agent_cluster_id in order detect and
count usage across contexts. This CL focuses on returning the agent
and the next will use that information for a use counter.

Bug: 1224926
Change-Id: I247d362b2ccef46cff64ad035a6aa8545612ccb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3043457
Commit-Queue: Ari Chivukula <arichiv@chromium.org>
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Chris Palmer <palmer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#904602}
NOKEYCHECK=True
GitOrigin-RevId: b414f982e66e8c84d05fe2be1d140dd0769d5b87