[GridNG] Unification of the paths of grid's out of flow items
Previously, the containing block of a grid out of flow item would be
computed and cached in |NGLogicalOutOfFlowPositionedNode|. This however,
forced us to have a way of re-computing the containing block rect in
|NGSimplifiedLayoutAlgorithm| when out of flow items change their
position in the grid dynamically. In this change, we directly compute
the containing block rect in |NGOutOfFlowLayoutPart|, which simplifies:
the need for |PlaceOutOfFlowDescendants| in |NGGridLayoutAlgorithm|, and
the caching and computation of the containing block rect in
|NGLogicalOutOfFlowPositionedNode| and |NGSimplifiedLayoutAlgorithm|
respectively.

Bug: 1045599
Change-Id: Ib2bd800ee03f17a02996d3b110ce7bdf22280448
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2917838
Commit-Queue: Ana Sollano Kim <ansollan@microsoft.com>
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#904860}
NOKEYCHECK=True
GitOrigin-RevId: db479920b31c682c1b3bd7c7b14263e8d21ff7b7