Add debug id parameter to crashReportPrivate.reportError
Add debugId as an optional parameter in crashReportPrivate.reportError.

When source maps are uploaded into the backend, they are identified by
URL and debug_id -- the debug_id is to distinguish between different
versions of the source maps. The same debug_id used to upload the source
map then needs to passed along with the error report so that the correct
source map can be found in the source map database.

BUG=b:186889239

Change-Id: I603b6a8ec45e3a14c5c03fccc8800abaa09a22b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3044846
Reviewed-by: Darren Shen <shend@chromium.org>
Reviewed-by: Ben Wells <benwells@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#904957}
NOKEYCHECK=True
GitOrigin-RevId: 18d687f897dfc6b38bf0fd0ceb800bea53acc0c6