Implement symmetric attribute
This is the second patch of a series of CL to implement MathML stretchy
operators. It adds step 4 from [1] to the operator algorithm in order
to ensure that some vertical operators (e.g. fences) are stretched
symmetrically above and below the math axis.

[1] https://w3c.github.io/mathml-core/#layout-of-operators

Bug: 6606, 1124301
Change-Id: I1a75f03a946d0781ce73f52f99b70c871b2dfeed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3013795
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#905033}
NOKEYCHECK=True
GitOrigin-RevId: 1b2791131aabba6b11c484fcfc666a60717f2bb5