Send storage key with CommitNavigationParams
This CL adds the storage key of the to-be-committed document to the
CommitNavigationParams, so that the browser can propagate to the
renderer the document's storage key.

Notice that since the document's origin is anyway computed by blink at
the moment, the storage key sent in the CommitNavigationParams is
provisional, and its origin is replaced by the correct one in blink.

Bug: 1224901
Change-Id: Id6eeed89dc109e5ea0701feb81074fff6f0a3de0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3003365
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/master@{#905178}
NOKEYCHECK=True
GitOrigin-RevId: 3e8de6d7d7261134a51661e392993f6cf7741203