bindings: Fix RejectedPromises not to use an invalid context.
RejetedPromises::Message::HasPromise is called even when
`script_state_` is invalidated, so it's wrong to enter the
context in HasPromise.  Plus, there is no need to enter the
context.  This patch fixes the wrong use of the invalid
context.

crbug.com/1224330 may be related to this problem, but no
confidence.  Hope this fixes the issue.

Bug: 1224330
Change-Id: Ib4660401b80341e3cbb9db807b06225e4e52fd1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3054237
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#905287}
NOKEYCHECK=True
GitOrigin-RevId: ac748b2b465ce292f5843578fe234a414d633a5e