Delay context lost event for canvas context in iframe
With the way that contextlost events are currently fired, contexts
in main frame and in iframe (from different domain) could receive
context loss event for a single user at exactly the same time (this
has been tested locally). With this structure, the script could
correlate the user's identity in both domains.

This CL delays contextlost events for canvas in iframe by a random delay
less than 100ms. It is not necessary to do this for contextrestored,
as contextrestored events will be delayed as well by delaying the
initial contextlost.

I have confirmed this is working locally as a test would be quite
difficult to implement.

Bug: 1164059

Change-Id: I2e9e811c4e8087b9d56b031d8b49d0b46945b026
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3043119
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Yi Xu <yiyix@chromium.org>
Cr-Commit-Position: refs/heads/master@{#905338}
NOKEYCHECK=True
GitOrigin-RevId: ae932ecff4ec7e0da28ff2c48f6c2e9f094c5839