Release dropped events timer before shutdown
This timer seems to keep the WidgetInputHandlerManager class alive as
it's a RefCountedThreadSafe and the timer callback binds to |this|. This
might cause a CHECK failure once the timer callback finishes as there is
some work done during destruction. See stack trace in bug for more
details.

Bug: 1233003
Change-Id: Id8f1ef68a62108f3d86a2bbfebd21b2eecaba299
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3053743
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#905360}
NOKEYCHECK=True
GitOrigin-RevId: 511d2dffbcb26d2c8fd721368c873bc73ff64d14