WebGPU: Require usage of the WebGPUInterface to check the context
This commit is in preparation for a future change where the context
provider may be destroyed as soon as the execution context is
destroyed, or if there is a context lost event.

If the context provider is destroyed, it is not safe to use the
WebGPUInterface. Remove all getters that return a raw pointer to
the interface, and change all callsites to first check if the
context provider is alive before accessing it.

Bug: 1231599
Change-Id: I59a9b8b820cfce82132018b4fce963c73e1cac32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3048517
Commit-Queue: Austin Eng <enga@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#905410}
NOKEYCHECK=True
GitOrigin-RevId: 9935983d3b4e41bad2bf98a2883a3439808bb042