Implement minsize/maxsize attributes
This is the third patch of a series of CL to implement MathML stretchy
operators. It implements step 5 from [1] so that users can specify some
minsize/maxsize values for vertical operators. It also adds missing
invalidation of minsize/maxsize when parsing them on the DOM side.

This CL also fixes the test mo-minsize-maxsize-001.html when checking
minsize with target size is 0. The test seemed to assume the minsize
would be double when applying the symmetric property but this is not
what is currently described in the spec.

[1] https://w3c.github.io/mathml-core/#layout-of-operators

Bug: 6606, 1124301
Change-Id: I4358ea65b0d74667ee6c5d8b2a11c6403f86e210
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3013895
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#905460}
NOKEYCHECK=True
GitOrigin-RevId: 3f72530b126ecd9c6e6a016c2249a1062d52f17d