[LayoutNG] OOF inline offset with inline containing block
When computing the inline containing block geometry in
ComputeInlineContainerGeometryForFragmentainer(), we had originally
calculated the offsets so that they were relative to the top of the
fragmentation context root. However, we weren't taking the inline
offset of the containing block into account, which can lead to
the wrong OOF offset.

Instead, calculate the offsets so that they are relative to the
containing block and add the offset of the containing block relative
to the fragmentation context root during the final inline CB rect
offset calculation. This ensures that we are also adding in the
inline offset from the containing block.

This update also resulted in a simplification of the number of input
variables and calculations previously needed in
ComputeInlineContainerGeometryForFragmentainer().

Bug: 1231218
Change-Id: I6d5bd2bffab0fde0a7b29b405494b33273eba2f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3042100
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#905485}
NOKEYCHECK=True
GitOrigin-RevId: 9b56cd568584faf7bb73a20e588b2b9b38450fdf