Add system fallback font UMA
Following r895250 crrev.com/c/2975052 which added UMA for
PrimaryFont, this patch adds UMA for the system fallback font.

The hypothesis is that the system fallback occurs much less
often than the primary font, but it is much longer when it
occurs. There are multiple ideas to improve this. Adding UMA
should help us to diagnose actual data and improvements as we
experiment ideas such as deferred shaping[1].

[1] https://docs.google.com/document/d/1ELdqjgAoh1Y_UYrnAWle6Ny8_QiWH9o2LBnTNEtB4OY/edit?usp=sharing

Bug: 1006253
Change-Id: Ic6a6e7f261623593e7cc14261c92c7daac9b59ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3054750
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#905921}
NOKEYCHECK=True
GitOrigin-RevId: b688f9194e33d11993960ce07c9fca1e6c01423e