Attempt to de-flake simulcast WPTs.
I believe the reason for flakily timing out is that ramp-up does not
happen quickly enough. Because I am unable to change the getUserMedia()
resolution in external/wpt/webrtc/simulcast/simulcast.js (this causes
timeout, either because the camera resolution is not supported by the
fake capturer or because webrtc does not send QQVGA, not sure?) this CL
changes the simulcast tests only to do 2 layers instead of 3.

2 layers instead of 3 should have a significantly lower ramp-up time in
cases where ramp-up is not instant.

The tests are already marked as having a long timeout so not sure how
else to de-flake other than to reduce the bitrate.

If any of the tests starts flaking again we should mark them all as
flaky and file a common bug for all tests instead of having one crbug
per test.

Bug: chromium:1223773, chromium:1225643, chromium:1229785, chromium:1230534
Change-Id: I7a78ffc63a0521e4f971611d28c283c512371f43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3057576
Commit-Queue: Henrik Boström <hbos@chromium.org>
Reviewed-by: Philipp Hancke <philipp.hancke@googlemail.com>
Cr-Commit-Position: refs/heads/master@{#906200}
NOKEYCHECK=True
GitOrigin-RevId: e38411c4731e03e0cf89d8d6e1300e731fcf4f7e