Revert "Manually clamp on suspected hot paths"
This reverts commit fc1469c6b11e2860bad1fa9740193ad4df813bfb.

Reason for revert:

This looks like it was not needed. It turns out that the massive regression in the original bug was actually a completely different change that broke tons of tests. All the regressed metrics look like they recovered in r905171 – r905193, and that includes https://crrev.com/c/3053750 which directly calls out that it was causing regressions.

Given that that seems to be the cause for the regressions, reverting these, as they’re probably fine. (If things regress with this reversion, I’ll re-commit, but it’s probably fine.)

Original change's description:
> Manually clamp on suspected hot paths
>
> https://crrev.com/c/3045786 seems to have caused a layout speed
> regression; manually clamp in two suspected hot paths.
>
> Bug: 1232666
> Change-Id: I08b0b24618f2a2c2e9b5b9f27b81b23ef5f7042b
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3054713
> Reviewed-by: Justin Novosad <junov@chromium.org>
> Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
> Commit-Queue: Avi Drissman <avi@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#905468}

Bug: 1232666
Change-Id: I66e14014dd57c761ce0a134d31e5ba5851b89ce3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3057651
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#906445}
NOKEYCHECK=True
GitOrigin-RevId: e244241a0d08b0a9f555f1dab50b7b8c1c599d65