Fix fuzzy matching on dynamic-content-change-rendering.html
I added fuzzy matching to this test in https://crrev.com/905947 but that
caused the test to fail everywhere [1] due to requiring a certain number
of pixels to be different. This patch fixes that so the test now passes.

I confirmed locally using the WPT runner that the test failed before
https://crrev.com/905947 and now passes. I am unable to reproduce the
original pixel differences seen on wpt.fyi so this is still a
speculative fix.

[1] https://wpt.fyi/results/html/semantics/forms/the-option-element/dynamic-content-change-rendering.html?label=master&label=experimental&aligned&q=dynamic-content-change-rendering.html

Change-Id: I8ea0572fee2b84d518df2daff51742155f3b4846
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3059121
Auto-Submit: Philip Rogers <pdr@chromium.org>
Commit-Queue: Kent Tamura <tkent@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#906448}
NOKEYCHECK=True
GitOrigin-RevId: d3bd39861c9ed398f857c989f642e3b844ea9474